Update timings for SR2017
This brings the start of the knockouts forwards a little in order
to ensure that there is enough time for them without substantially
overrunning.

We also introduce an intentional delay at the time we expect to
switch to only using a single arena, in order to provide time for
the actual change-over to happen.

Change-Id: I4b2c77e404355bb794268c5172ef69ea9456165c